Move generic XML reader & parser to their own subproject 'xml-reader'
