--- map-reader/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-    implementation project(":map-data")
-}
