+++ settings.gradle
@@ -15,3 +15,4 @@ include 'maps-server'
 include 'smoke-testing'
 include 'swing-lib'
 include 'test-common'
+include 'xml-reader'
