+++ xml-reader/build.gradle
@@ -1,4 +1,3 @@
 dependencies {
     implementation project(":java-extras")
-    implementation project(":xml-reader")
 }
