+++ xml-reader/src/main/java/org/triplea/map/reader/generic/xml/Tag.java
@@ -4,5 +4,4 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Tag {
-}
+public @interface Tag {}
