+++ xml-reader/src/main/java/org/triplea/map/reader/generic/xml/XmlParser.java
